Continue RPC implementation
