+++ src/containers/src/container_main.cpp
@@ -1,4 +1,4 @@
-#include "container_rpc.hpp"
+#include "include/container_rpc.hpp"
 
 int main(int argc, char* argv[]) {
   clipper::container::RPC rpc;
